Split out AndroidStorage.cpp from android/jni/app-android.cpp, move to Common
